% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{congeneric_model}
\alias{congeneric_model}
\title{Specify a congeneric measurement model}
\usage{
congeneric_model(lambdaA, lambdaB, thetaA = NULL, thetaB = NULL)
}
\arguments{
\item{lambdaA, lambdaB}{Numeric vectors of factor loadings for scale A and
scale B.}

\item{thetaA, thetaB}{Optional numeric vectors of residual variances for scale A
and scale B. If supplied, each must have the same length as the corresponding
loading vector.}
}
\value{
An object of class \code{"latentoverlap_spec"} for use in \code{\link[=rho_correction]{rho_correction()}}.
}
\description{
Creates a measurement-model specification in which item loadings may differ
across items within each scale. This is the most flexible measurement-model
specification currently supported by the correction function.
}
\examples{
congeneric_model(
  lambdaA = c(.70, .75, .80, .85, .90),
  lambdaB = c(.65, .70, .75, .80),
  thetaA = 1 - c(.70, .75, .80, .85, .90)^2,
  thetaB = 1 - c(.65, .70, .75, .80)^2
)

}
