% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{parallel_model}
\alias{parallel_model}
\title{Specify a parallel measurement model}
\usage{
parallel_model(
  lambdaA,
  lambdaB,
  kA = NULL,
  kB = NULL,
  thetaA = NULL,
  thetaB = NULL
)
}
\arguments{
\item{lambdaA, lambdaB}{Factor loading(s) for scale A and scale B. Each may be
a scalar or a vector of length \code{kA} / \code{kB}. If unequal vectors are supplied,
their mean is used to enforce the parallel model.}

\item{kA, kB}{Number of items in scale A and scale B. Required when the
corresponding loading and residual variance are supplied as scalars.}

\item{thetaA, thetaB}{Optional residual variance(s) for scale A and scale B.
If \code{NULL}, residual variances are set to \code{1 - lambda^2}. If unequal vectors
are supplied, their mean is used to enforce the parallel model.}
}
\value{
An object of class \code{"latentoverlap_spec"} for use in \code{\link[=rho_correction]{rho_correction()}}.
}
\description{
Creates a measurement-model specification in which items within each scale
have equal factor loadings and equal residual variances.
}
\examples{
parallel_model(
  lambdaA = .80,
  lambdaB = .75,
  kA = 6,
  kB = 5
)

}
