% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{tau_model}
\alias{tau_model}
\title{Specify a tau-equivalent measurement model}
\usage{
tau_model(lambdaA, lambdaB, kA = NULL, kB = NULL, thetaA = NULL, thetaB = NULL)
}
\arguments{
\item{lambdaA, lambdaB}{Factor loading(s) for scale A and scale B. Each may be
a scalar or a vector of length \code{kA} / \code{kB}. If unequal vectors are supplied,
their mean is used to enforce tau-equivalence.}

\item{kA, kB}{Number of items in scale A and scale B. Required when the
corresponding loading and residual variance are supplied as scalars.}

\item{thetaA, thetaB}{Optional residual variance(s) for scale A and scale B.
May be \code{NULL}, scalar, or a vector of length \code{kA} / \code{kB}.}
}
\value{
An object of class \code{"latentoverlap_spec"} for use in \code{\link[=rho_correction]{rho_correction()}}.
}
\description{
Creates a measurement-model specification in which items within each scale
have equal factor loadings. Residual variances may differ across items if
supplied as vectors.
}
\examples{
tau_model(
  lambdaA = .80,
  lambdaB = .75,
  kA = 6,
  kB = 5,
  thetaA = .36,
  thetaB = .44
)

}
